+++ js/bmf-buildform.js
@@ -2,7 +2,6 @@
 (function($){
     // register as jquery function
     $.fn.bmf_buildform = function(){
-        $(this).bmf_inlineform();
         $(this).bmf_autocomplete();
         $(this).bmf_calendar();
     };