+++ tests/formapi/tests.py
@@ -6,15 +6,14 @@ from __future__ import unicode_literals
 
 from django.contrib.contenttypes.models import ContentType
 
-from djangobmf.utils import get_model_from_cfg
 from djangobmf.testcase import BMFModuleTestCase
 
 
-class FormTests(BMFModuleTestCase):
-
-    def test_forms(self):
-        """
-        """
+#class FormTests(BMFModuleTestCase):
+#
+#    def test_forms(self):
+#        """
+#        """
 #       self.model = get_model_from_cfg("POSITION")
 
 #       # don't accept get