+++ CHANGELOG.rst
@@ -1,11 +1,22 @@
+Version 0.2.1 (currently in development)
+========================================
+
+* Added docker Support
+* New forms which use django-forms and no inline formsets
+* Changed Admin-API urls
+* Added Celery
+* Added the function to test a single module or only the core
+* Updates in ``contrib.accounting``
+* Updates in ``contrib.quotation``
+
 Version 0.2.0
-==================
+========================================
 
 * Redesigned notifications
 * All contrib models are ``swappable``
 * Initialised ``contrib.timesheets``
 
 Version 0.1.4
-==================
+========================================
 
 * First public release